addition of ntfsvolume utility
Utility reads volume-info without parsing the MFT, which can be
time consuming. The behavior is exactly like that of ntfscluster
when invoked without --file or --inode arguments except that the
MFT is not scanned.

Correspondingly, the output of ntfsvolume does not include
used / total / metadata statistics. Otherwise, the content of
ntfsvolume's output is exactly the same.